fix: fallback to endpoints if endpointslice unavailable
